+++ library/src/schemas/enumType/types.ts
@@ -1,6 +1,0 @@
-/**
- * Enum type.
- */
-export type Enum<TOption extends string = string> =
-  | Readonly<[TOption, ...TOption[]]>
-  | [TOption, ...TOption[]];