+++ library/CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to the library will be documented in this file.
 - Change type of `array`, `arrayAsync`, `intersection`, `intersectionAsync`, `map`, `mapAsync`, `object`, `objectAsync`, `union`, `unionAsync`, `record`, `recordAsync`, `set`, `setAsync`, `tuple` and `tupleAsync` schema
 - Rename `schema` property of every schema type to `type`
 - Rename `intersection` and `intersectionAsync` schema to `intersect` and `intersectAsync`
+- Rename `enumType` and `enumTypeAsync` schema to `picklist` and `picklistAsync`
 - Rename `default` property of `optional`, `optionalAsync`, `nullable`, `nullableAsync`, `nullish` and `nullishAsync` schema to `getDefault`
 - Rename `ObjectShape` and `ObjectShapeAsync` types to `ObjectEntries` and `ObjectEntriesAsync`
 - Rename `TupleShape` and `TupleShapeAsync` types to `TupleItems` and `TupleItemsAsync`